DeepSESRApi/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.3
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
keras==3.6.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pillow==11.0.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-multipart==0.0.12
requests==2.32.3
rich==13.9.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
opencv-python==4.8.1.78
moviepy==1.0.3
aiofiles==23.2.1